requirements.txt
bidict==0.22.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
gTTS==2.4.0
h11==0.14.0
idna==3.4
mpg123==0.4
python-decouple==3.8
python-engineio==4.8.0
python-socketio==5.10.0
pyttsx3==2.90
requests==2.31.0
simple-websocket==1.0.0
tk==0.1.0
urllib3==2.0.7
wsproto==1.2.0
